crank up the warning levels for GCC builds, to better approximate the warnings that Flash and AIR have enabled. Then, fix those warnings.
Note: when building on GCC 4.3 or later we now enable warnings-as-errors, as that version of gcc has added a way to disable the "clobbered" warning we've been living with.

+++ core/AbcParser.cpp
@@ -632,7 +632,7 @@ namespace avmplus
 				core->console << "        name_index=" << name_index;
 				if (name_index > 0 && name_index < pool->constantStringCount)
 					core->console << " \"" << pool->cpool_string[name_index] << "\"";
-				core->console << "\n        flags=" << abcFlags << "\n";
+				core->console << "\n        flags=" << (uint32_t)abcFlags << "\n";
 			)
 
 			const NativeMethodInfo* ni = NULL;
